Chore: restructured page folders
